Refactor/overall (#69)
* Refactored parsers

* Refactored file names, relocated Lecture class

* Other minor refactor changes, fixed no internet reloading

* Fixed database concurrency issue
