+++ app_feup/lib/Main.dart
@@ -1,4 +1,5 @@
 import 'package:app_feup/model/SchedulePageModel.dart';
+import 'package:app_feup/model/entities/Exam.dart';
 import 'package:app_feup/view/Pages/ClassificationsPageView.dart';
 import 'package:app_feup/view/Pages/ExamsPageView.dart';
 import 'package:app_feup/view/Pages/HomePageView.dart';
@@ -14,9 +15,8 @@ import 'package:flutter_redux/flutter_redux.dart';
 import 'view/Theme.dart';
 import 'model/AppState.dart';
 import 'package:redux/redux.dart';
-import 'redux/reducers.dart';
-import 'controller/parsers/parser-exams.dart';
-import 'package:app_feup/redux/actionCreators.dart';
+import 'redux/Reducers.dart';
+import 'package:app_feup/redux/ActionCreators.dart';
 import 'package:app_feup/model/AppState.dart';
 import 'package:app_feup/controller/LifecycleEventHandler.dart';
