+++ app_feup/lib/controller/LoadInfo.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 import 'package:app_feup/controller/local_storage/AppSharedPreferences.dart';
-import 'package:app_feup/redux/actionCreators.dart';
+import 'package:app_feup/redux/ActionCreators.dart';
 import 'package:app_feup/redux/RefreshItemsAction.dart';
-import 'package:app_feup/redux/actions.dart';
+import 'package:app_feup/redux/Actions.dart';
 import 'package:tuple/tuple.dart';
 import 'package:app_feup/model/AppState.dart';
 import 'package:redux/redux.dart';
@@ -26,17 +26,15 @@ Future loadRemoteUserInfoToState(Store<AppState> store){
       exams = new Completer(),
       schedule = new Completer(),
       printBalance = new Completer(),
-      feesBalance = new Completer(),
-      feesLimit = new Completer(),
+      fees = new Completer(),
       coursesStates = new Completer();
   store.dispatch(getUserInfo(userInfo));
   store.dispatch(getUserExams(exams));
   store.dispatch(getUserSchedule(schedule));
   store.dispatch(getUserPrintBalance(printBalance));
-  store.dispatch(getUserFeesBalance(feesBalance));
-  store.dispatch(getUserFeesNextLimit(feesLimit));
+  store.dispatch(getUserFees(fees));
   store.dispatch(getUserCoursesState(coursesStates));
-  return Future.wait([exams.future, schedule.future, printBalance.future, feesBalance.future, feesLimit.future, coursesStates.future, userInfo.future]);
+  return Future.wait([exams.future, schedule.future, printBalance.future, fees.future, coursesStates.future, userInfo.future]);
 }
 
 void loadLocalUserInfoToState(store) async {
