+++ app_feup/lib/controller/local_storage/AppDatabase.dart
@@ -2,11 +2,13 @@ import 'dart:async';
 import 'package:sqflite/sqflite.dart';
 import 'dart:io';
 import 'package:path_provider/path_provider.dart';
+import 'package:synchronized/synchronized.dart';
 
 class AppDatabase {
   Database _db;
   String name;
   String command;
+  static Lock lock = new Lock();
 
   AppDatabase(String name, String command) {
     this.name = name;
@@ -17,6 +19,16 @@ class AppDatabase {
     if (_db == null)
       _db = await initializeDatabase();
     return _db;
+  }
+
+  insertInDatabase(String table, Map<String, dynamic> values,
+      {String nullColumnHack, ConflictAlgorithm conflictAlgorithm}) async{
+
+      lock.synchronized(() async {
+        Database db = await getDatabase();
+
+        db.insert(table, values, nullColumnHack: nullColumnHack, conflictAlgorithm: conflictAlgorithm);
+      });
   }
 
   Future<Database> initializeDatabase() async {
