+++ app_feup/lib/controller/parsers/ParserCourses.dart
@@ -1,16 +1,9 @@
 import 'package:html/parser.dart' show parse;
+import 'package:http/http.dart' as http;
 import 'dart:async';
-import 'package:app_feup/controller/networking/NetworkRouter.dart';
-import 'package:app_feup/model/AppState.dart';
-import 'package:redux/redux.dart';
 import 'dart:collection';
 
-Future<Map<String, String>> getCourseState(String link, Store<AppState> store) async{
-  Map<String, String> query = {"pv_num_unico": store.state.content['session'].studentNumber};
-
-  String cookies = store.state.content['session'].cookies;
-
-  var response = await NetworkRouter.getWithCookies(link, query, cookies);
+Future<Map<String, String>> parseCourses(http.Response response) async{
 
   var document = await parse(response.body);
