+++ app_feup/lib/model/LoginPageModel.dart
@@ -2,7 +2,7 @@ import 'package:app_feup/model/AppState.dart';
 import 'package:app_feup/view/Pages/LoginPageView.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_redux/flutter_redux.dart';
-import 'package:app_feup/redux/actionCreators.dart';
+import 'package:app_feup/redux/ActionCreators.dart';
 import 'package:redux/redux.dart';
 
 class LoginPage extends StatefulWidget {
