+++ app_feup/lib/redux/ActionCreators.dart
@@ -1,19 +1,21 @@
 import 'dart:async';
-import 'package:app_feup/controller/loadinfo.dart';
+import 'package:app_feup/controller/LoadInfo.dart';
 import 'package:app_feup/controller/local_storage/AppExamsDatabase.dart';
 import 'package:app_feup/controller/local_storage/AppLecturesDatabase.dart';
 import 'package:app_feup/controller/local_storage/AppSharedPreferences.dart';
-import 'package:app_feup/controller/parsers/parser-exams.dart';
-import 'package:app_feup/controller/parsers/parser-schedule.dart';
-import 'package:app_feup/controller/parsers/parser-prints.dart';
-import 'package:app_feup/controller/parsers/parser-fees.dart';
-import 'package:app_feup/controller/parsers/parser-courses.dart';
+import 'package:app_feup/controller/parsers/ParserExams.dart';
+import 'package:app_feup/controller/parsers/ParserSchedule.dart';
+import 'package:app_feup/controller/parsers/ParserPrintBalance.dart';
+import 'package:app_feup/controller/parsers/ParserFees.dart';
+import 'package:app_feup/controller/parsers/ParserCourses.dart';
 import 'package:app_feup/model/entities/CourseUnit.dart';
+import 'package:app_feup/model/entities/Exam.dart';
+import 'package:app_feup/model/entities/Lecture.dart';
 import 'package:app_feup/model/entities/Session.dart';
 import 'package:redux_thunk/redux_thunk.dart';
 import 'package:tuple/tuple.dart';
 import '../model/AppState.dart';
-import 'actions.dart';
+import 'Actions.dart';
 import 'package:redux/redux.dart';
 import 'package:app_feup/controller/networking/NetworkRouter.dart';
 
@@ -62,10 +64,10 @@ ThunkAction<AppState> getUserInfo(Completer<Null> action) {
       final profile = NetworkRouter.getProfile(store.state.content['session']).then((res) => store.dispatch(new SaveProfileAction(res)));
       final ucs = NetworkRouter.getCurrentCourseUnits(store.state.content['session']).then((res) => store.dispatch(new SaveUcsAction(res)));
       await Future.wait([profile, ucs]);
-      action.complete();
     } catch (e) {
-      print(e);
+      print("Failed to get User Info");
     }
+    action.complete();
   };
 }
 
@@ -91,7 +93,10 @@ ThunkAction<AppState> getUserExams(Completer<Null> action) {
       //need to get student course here
       store.dispatch(new SetExamsStatusAction(RequestStatus.BUSY));
 
-      List<Exam> courseExams = await examsGet("https://sigarra.up.pt/${store.state.content['session'].faculty}/pt/exa_geral.mapa_de_exames?p_curso_id=742");
+      List<Exam> courseExams = await parseExams(
+          await NetworkRouter.getWithCookies(NetworkRouter.getBaseUrlFromSession(store.state.content['session']) + "exa_geral.mapa_de_exames?p_curso_id=742",
+          {}, store.state.content['session'].cookies)
+      );
 
       List<CourseUnit> userUcs = store.state.content['currUcs'];
       List<Exam> exams = new List<Exam>();
@@ -106,7 +111,6 @@ ThunkAction<AppState> getUserExams(Completer<Null> action) {
 
         }
       }
-      action.complete();
 
       // Updates local database according to the information fetched -- Exams
       Tuple2<String, String> userPersistentInfo = await AppSharedPreferences.getPersistentUserInfo();
@@ -118,8 +122,11 @@ ThunkAction<AppState> getUserExams(Completer<Null> action) {
       store.dispatch(new SetExamsAction(exams));
       
     } catch (e) {
+      print("Failed to get Exams");
       store.dispatch(new SetExamsStatusAction(RequestStatus.FAILED));
     }
+
+    action.complete();
   };
 }
 
@@ -137,17 +144,12 @@ ThunkAction<AppState> getUserSchedule(Completer<Null> action) {
           date.month.toString().padLeft(2, '0') +
           date.day.toString().padLeft(2, '0');
 
-      List<Lecture> lectures = await scheduleGet(
+      List<Lecture> lectures = await parseSchedule(
           await NetworkRouter.getWithCookies(
-              "https://sigarra.up.pt/${store.state.content['session']
-                  .faculty}/pt/mob_hor_geral.estudante?pv_codigo=${store.state
-                  .content['session']
-                  .studentNumber}&pv_semana_ini=$beginWeek&pv_semana_fim=$endWeek",
+              NetworkRouter.getBaseUrlFromSession(store.state.content['session'])
+                  + "mob_hor_geral.estudante?pv_codigo=${store.state.content['session'].studentNumber}"
+                  "&pv_semana_ini=$beginWeek&pv_semana_fim=$endWeek",
               {}, store.state.content['session'].cookies));
-
-      action.complete();
-
-
 
       // Updates local database according to the information fetched -- Lectures
       Tuple2<String, String> userPersistentInfo = await AppSharedPreferences.getPersistentUserInfo();
@@ -159,8 +161,10 @@ ThunkAction<AppState> getUserSchedule(Completer<Null> action) {
       store.dispatch(new SetScheduleAction(lectures));
       store.dispatch(new SetScheduleStatusAction(RequestStatus.SUCCESSFUL));
     } catch (e) {
+      print("Failed to get Schedule");
       store.dispatch(new SetScheduleStatusAction(RequestStatus.FAILED));
     }
+    action.complete();
   };
 }
 
@@ -173,44 +177,68 @@ ThunkAction<AppState> updateSelectedPage(new_page) {
 ThunkAction<AppState> getUserPrintBalance(Completer<Null> action) {
   return (Store<AppState> store) async {
 
-    String url = "https://sigarra.up.pt/${store.state.content['session'].faculty}/pt/imp4_impressoes.atribs?";
+    String url = NetworkRouter.getBaseUrlFromSession(store.state.content['session']) + "imp4_impressoes.atribs?";
 
-    String printBalance = await getPrintsBalance(url, store);
+    Map<String, String> query = {"p_codigo": store.state.content['session'].studentNumber};
+
+    String cookies = store.state.content['session'].cookies;
+
+    try {
+      var response = await NetworkRouter.getWithCookies(url, query, cookies);
+      String printBalance = await getPrintsBalance(response);
+      store.dispatch(new SetPrintBalanceAction(printBalance));
+    }catch(e){
+      print("Failed to get Print Balance");
+    }
     action.complete();
-    store.dispatch(new SetPrintBalanceAction(printBalance));
-
   };
 }
 
-ThunkAction<AppState> getUserFeesBalance(Completer<Null> action) {
+ThunkAction<AppState> getUserFees(Completer<Null> action) {
   return (Store<AppState> store) async {
 
-    String url = "https://sigarra.up.pt/${store.state.content['session'].faculty}/pt/gpag_ccorrente_geral.conta_corrente_view?";
+    String url = NetworkRouter.getBaseUrlFromSession(store.state.content['session']) + "gpag_ccorrente_geral.conta_corrente_view?";
 
-    String feesBalance = await getFeesBalance(url, store);
+    Map<String, String> query = {"pct_cod": store.state.content['session'].studentNumber};
+
+    String cookies = store.state.content['session'].cookies;
+
+    try{
+      var response = await NetworkRouter.getWithCookies(url, query, cookies);
+
+      String feesBalance = await parseFeesBalance(response);
+      store.dispatch(new SetFeesBalanceAction(feesBalance));
+
+      String feesLimit = await parseFeesNextLimit(response);
+      store.dispatch(new SetFeesLimitAction(feesLimit));
+    }catch(e){
+      print("Failed to get Fees info");
+    }
+
     action.complete();
-    store.dispatch(new SetFeesBalanceAction(feesBalance));
-  };
-}
-
-ThunkAction<AppState> getUserFeesNextLimit(Completer<Null> action) {
-  return (Store<AppState> store) async {
-
-    String url = "https://sigarra.up.pt/${store.state.content['session'].faculty}/pt/gpag_ccorrente_geral.conta_corrente_view?";
-
-    String feesLimit = await getFeesNextLimit(url, store);
-    action.complete();
-    store.dispatch(new SetFeesLimitAction(feesLimit));
   };
 }
 
 ThunkAction<AppState> getUserCoursesState(Completer<Null> action) {
   return (Store<AppState> store) async {
 
-    String url = "https://sigarra.up.pt/${store.state.content['session'].faculty}/pt/fest_geral.cursos_list?";
+    String url = NetworkRouter.getBaseUrlFromSession(store.state.content['session']) + "fest_geral.cursos_list?";
 
-    Map<String,String> coursesStates = await getCourseState(url, store);
+    Map<String, String> query = {"pv_num_unico": store.state.content['session'].studentNumber};
+
+    String cookies = store.state.content['session'].cookies;
+
+    try{
+      var response = await NetworkRouter.getWithCookies(url, query, cookies);
+
+      Map<String,String> coursesStates = await parseCourses(response);
+
+      store.dispatch(new SetCoursesStatesAction(coursesStates));
+
+    }catch(e){
+      print("Failed to get Fees info");
+    }
+
     action.complete();
-    store.dispatch(new SetCoursesStatesAction(coursesStates));
   };
 }
