+++ app_feup/lib/redux/Reducers.dart
@@ -1,5 +1,5 @@
 import '../model/AppState.dart';
-import 'actions.dart';
+import 'Actions.dart';
 
 AppState appReducers(AppState state, dynamic action) {
   if (action is SaveLoginDataAction) {
