--- app_feup/lib/tests/example.dart
@@ -1,8 +1,0 @@
-import 'package:flutter_test/flutter_test.dart';
-
-void main() {
-  test('my first unit test', () {
-    var answer = 42;
-    expect(answer, 42);
-  });
-}
