+++ app_feup/lib/view/Pages/GeneralPageView.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import '../widgets/NavigationDrawer.dart';
+import '../Widgets/NavigationDrawer.dart';
 import 'package:app_feup/model/AppState.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:app_feup/controller/loadinfo.dart';
+import 'package:app_feup/controller/LoadInfo.dart';
 import 'package:app_feup/model/ProfilePageModel.dart';
 
 abstract class GeneralPageView extends StatelessWidget {
