+++ app_feup/lib/view/Widgets/DateRectangle.dart
