+++ app_feup/lib/view/Widgets/ScheduleCard.dart
@@ -1,8 +1,8 @@
-import 'package:app_feup/controller/parsers/parser-schedule.dart';
 import 'package:app_feup/model/AppState.dart';
-import 'package:app_feup/view/widgets/DateRectangle.dart';
-import 'package:app_feup/view/widgets/GenericCard.dart';
-import 'package:app_feup/view/widgets/ScheduleRow.dart';
+import 'package:app_feup/model/entities/Lecture.dart';
+import 'package:app_feup/view/Widgets/DateRectangle.dart';
+import 'package:app_feup/view/Widgets/GenericCard.dart';
+import 'package:app_feup/view/Widgets/ScheduleRow.dart';
 import 'package:flutter_redux/flutter_redux.dart';
 import 'package:flutter/material.dart';
