+++ app_feup/lib/view/Widgets/ScheduleEventRectangle.dart
