+++ app_feup/lib/view/Widgets/ScheduleRow.dart
