+++ app_feup/lib/view/Widgets/ScheduleSlot.dart
