+++ app_feup/lib/view/Widgets/ScheduleTimeInterval.dart
